src/requirements/local.txt
-r ./base.txt

ipython
django-extensions
pylint
pycodestyle
ipdb